fix build exc

+++ app/src/main/java/cn/lancet/navigation/LancetExt.kt
@@ -13,4 +13,4 @@ val Float.px
 val Int.dp
    get() = TypedValue.applyDimension(
        TypedValue.COMPLEX_UNIT_DIP, this.toFloat(),Resources.getSystem().displayMetrics
-   )
+   ).toInt()
